# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MiniProyecto",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/taxes": {
            "get": {
                "tags": [
                    "Taxes"
                ],
                "summary": "Get taxes",
                "description": "Get list of taxes",
                "operationId": "getTaxes",
                "responses": {
                    "200": {
                        "description": "Todos los impuestos",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Taxes"
                ],
                "summary": "Create tax",
                "description": "Create and return one tax",
                "operationId": "create_tax",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "percentage",
                                    "enterprise_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "percentage": {
                                        "type": "integer"
                                    },
                                    "enterprise_id": {
                                        "format": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Pepe",
                                    "percentage": 12,
                                    "enterprise_id": 4
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Impuesto creado satisfactoriamente",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/taxes/find-by-enterprise/{enterpriseId}": {
            "get": {
                "tags": [
                    "Taxes"
                ],
                "summary": "Get taxes by enterrprise",
                "description": "Get taxes by enterprise id",
                "operationId": "getByEnterprise",
                "parameters": [
                    {
                        "name": "enterpriseId",
                        "in": "path",
                        "description": "Enterprise id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todos los impuestos",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/taxes/{taxId}": {
            "put": {
                "tags": [
                    "Taxes"
                ],
                "summary": "Update tax",
                "description": "Update and return one tax",
                "operationId": "update_tax",
                "parameters": [
                    {
                        "name": "taxId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "percentage",
                                    "enterprise_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "percentage": {
                                        "type": "integer"
                                    },
                                    "enterprise_id": {
                                        "format": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Pepe",
                                    "percentage": 12,
                                    "enterprise_id": 4
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Impuesto creado satisfactoriamente",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Taxes"
                ],
                "summary": "Delete tax",
                "description": "Delete one tax",
                "operationId": "delete_tax",
                "parameters": [
                    {
                        "name": "taxId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa registrada satisfactoriamente",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}